+++ R/get_lux.R
@@ -1,19 +1,3 @@
-.lux_scale <- c(
-  "NEO" = 1.25, "CLE" = 1.25,
-  "MOS0" = NA,
-  "MOS1" = 2.05, "MOS2" = 2.05,
-  "MOS3" = 2.67, "MOS4" = 2.67,
-  "MRA" = 3.25
-)
-
-.lux_max <- c(
-  "NEO" = 2500, "CLE" = 2500,
-  "MOS0" = NA,
-  "MOS1" = 5000, "MOS2" = 5000,
-  "MOS3" = 6500, "MOS4" = 6500,
-  "MRA" = 6000
-)
-
 #' Convert pre-parsed lux data into its final format
 #'
 #' @param lux output from \code{dev_parse_lux}, with minor modifications